requirements.txt
python-dotenv
discord.py
yt-dlp
pytube
pickledb
PyNaCl
youtube_dl